Remove unused constants and utils, move twitch hook to hooks
